Catch all possible migration errors

+++ database/migrations/2017_11_04_170844_changes_for_v470a.php
@@ -39,16 +39,18 @@ class ChangesForV470a extends Migration
      */
     public function down(): void
     {
-        try {
-            Schema::table(
-                'transactions',
-                static function (Blueprint $table) {
-                    $table->dropColumn('reconciled');
-                }
-            );
-        } catch (QueryException|ColumnDoesNotExist $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if (Schema::hasColumn('transactions', 'reconciled')) {
+            try {
+                Schema::table(
+                    'transactions',
+                    static function (Blueprint $table) {
+                        $table->dropColumn('reconciled');
+                    }
+                );
+            } catch (QueryException|ColumnDoesNotExist $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 
@@ -58,16 +60,18 @@ class ChangesForV470a extends Migration
      */
     public function up(): void
     {
-        try {
-            Schema::table(
-                'transactions',
-                static function (Blueprint $table) {
-                    $table->boolean('reconciled')->after('deleted_at')->default(0);
-                }
-            );
-        } catch (QueryException $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if (!Schema::hasColumn('transactions', 'reconciled')) {
+            try {
+                Schema::table(
+                    'transactions',
+                    static function (Blueprint $table) {
+                        $table->boolean('reconciled')->after('deleted_at')->default(0);
+                    }
+                );
+            } catch (QueryException $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 }
