+++ database/migrations/2018_04_07_210913_changes_for_v473.php
@@ -42,33 +42,36 @@ class ChangesForV473 extends Migration
      */
     public function down(): void
     {
-        try {
-            Schema::table(
-                'bills',
-                static function (Blueprint $table) {
-                    // cannot drop foreign keys in SQLite:
-                    if ('sqlite' !== config('database.default')) {
-                        $table->dropForeign('bills_transaction_currency_id_foreign');
+        if(!Schema::hasColumn('bills', 'transaction_currency_id')) {
+            try {
+                Schema::table(
+                    'bills',
+                    static function (Blueprint $table) {
+                        // cannot drop foreign keys in SQLite:
+                        if ('sqlite' !== config('database.default')) {
+                            $table->dropForeign('bills_transaction_currency_id_foreign');
+                        }
+                        $table->dropColumn('transaction_currency_id');
                     }
-                    $table->dropColumn('transaction_currency_id');
-                }
-            );
-        } catch (QueryException|ColumnDoesNotExist $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+                );
+            } catch (QueryException|ColumnDoesNotExist $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
 
-
-        try {
-            Schema::table(
-                'rules',
-                static function (Blueprint $table) {
-                    $table->dropColumn('strict');
-                }
-            );
-        } catch (QueryException|ColumnDoesNotExist $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if(!Schema::hasColumn('rules', 'strict')) {
+            try {
+                Schema::table(
+                    'rules',
+                    static function (Blueprint $table) {
+                        $table->dropColumn('strict');
+                    }
+                );
+            } catch (QueryException|ColumnDoesNotExist $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 
@@ -79,28 +82,32 @@ class ChangesForV473 extends Migration
      */
     public function up(): void
     {
-        try {
-            Schema::table(
-                'bills',
-                static function (Blueprint $table) {
-                    $table->integer('transaction_currency_id', false, true)->nullable()->after('user_id');
-                    $table->foreign('transaction_currency_id')->references('id')->on('transaction_currencies')->onDelete('set null');
-                }
-            );
-        } catch (QueryException $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if(!Schema::hasColumn('bills', 'transaction_currency_id')) {
+            try {
+                Schema::table(
+                    'bills',
+                    static function (Blueprint $table) {
+                        $table->integer('transaction_currency_id', false, true)->nullable()->after('user_id');
+                        $table->foreign('transaction_currency_id')->references('id')->on('transaction_currencies')->onDelete('set null');
+                    }
+                );
+            } catch (QueryException $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
-        try {
-            Schema::table(
-                'rules',
-                static function (Blueprint $table) {
-                    $table->boolean('strict')->default(true);
-                }
-            );
-        } catch (QueryException $e) {
-            Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
-            Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+        if(!Schema::hasColumn('rules', 'strict')) {
+            try {
+                Schema::table(
+                    'rules',
+                    static function (Blueprint $table) {
+                        $table->boolean('strict')->default(true);
+                    }
+                );
+            } catch (QueryException $e) {
+                Log::error(sprintf('Could not execute query: %s', $e->getMessage()));
+                Log::error('If the column or index already exists (see error), this is not an problem. Otherwise, please open a GitHub discussion.');
+            }
         }
     }
 }
